django-api/requirements.txt
Django
djangorestframework
django-cors-headers
googletrans==3.1.0a0
SpeechRecognition
requests
gTTS
gTTS-token